+++ src/main/java/hr/dlatecki/algorithms/gen_alg/annotations/package-info.java
@@ -1,0 +1,6 @@
+// Copyright (C) Domagoj Latečki
+// You may use, distribute and modify this code under the terms of the MIT license.
+/**
+ * Contains various annotations and a sub-package which defines annotation processing.
+ */
+package hr.dlatecki.algorithms.gen_alg.annotations;