+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/crossover/package-info.java
@@ -1,0 +1,9 @@
+// Copyright (C) Domagoj Latečki
+// You may use, distribute and modify this code under the terms of the MIT license.
+/**
+ * Contains implementations of crossover operators.
+ * 
+ * @author Domagoj Latečki
+ * @since 1.8
+ */
+package hr.dlatecki.algorithms.gen_alg.operators.crossover;