+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/mutation/GaussianDistributionMutation.java
@@ -19,7 +19,7 @@ import hr.dlatecki.algorithms.gen_alg.population.chromosomes.DoubleArrayChromoso
  * More info about the normal distribution can be found
  * <a href = "https://en.wikipedia.org/wiki/Normal_distribution">here</a>.<br>
  * <br>
- * This mutation can only be used on chromosomes which extend the <code>DoubleArrayChromosome</code>.
+ * This mutation operator can only be used on chromosomes which extend <code>DoubleArrayChromosome</code>.
  * 
  * @author Domagoj Latečki
  * @version 1.0